requirements.txt
face_recognition
opencv-python
numpy
PyYAML
